Update ComfyUI-Manager to desktop branch
Updates ComfyUI-Manager to the latest commit, and reverts the following PRs:

- https://github.com/Comfy-Org/ComfyUI-Manager/pull/2025
- Follow-up commit: https://github.com/Comfy-Org/ComfyUI-Manager/commit/4834874091f6e272d7e7d093d71cd383530a245f

This removes the requirement for every desktop user to download several new packages for the matrix share feature.  This is an inconvenience for users wanting to use matrix share, however including this update will cause major update pain for many international users with network issues.